Fix: RankingInternalController 타입 변환 누락 수정

+++ src/main/java/com/finlearn/rankingservice/presentation/RankingInternalController.java
@@ -28,7 +28,7 @@ public class RankingInternalController {
             @Valid @RequestBody UpdateRankingScoreRequest request
     ) {
         RankingEntryDto dto = rankingService.updateScore(
-                seasonId, request.getUserId(), request.getRankingType(), request.getScore());
+                seasonId, request.getUserId(), request.getRankingType(), request.getScore().doubleValue());
         return RankingEntryResponse.from(dto);
     }
